+++ packages/components/src/components/link/link.stories.ts
@@ -2,19 +2,26 @@
 import '../../solid-components';
 import { html } from 'lit-html';
 import { storybookDefaults, storybookHelpers, storybookTemplate } from '../../../scripts/storybook/helper';
-import { userEvent } from '@storybook/test';
-import { waitUntil } from '@open-wc/testing-helpers';
 import { withActions } from '@storybook/addon-actions/decorator';
 
 const { argTypes } = storybookDefaults('sd-link');
 const { generateTemplate } = storybookTemplate('sd-link');
 const { overrideArgs } = storybookHelpers('sd-link');
 
+/**
+ *
+ * Allows users to navigate to another location or perform actions like e.g. downloading files.
+ *
+ *  **Related templates**:
+ * - [Link List](?path=/docs/templates-link-list--docs)
+ */
+
 export default {
+  tags: ['!dev'],
   title: 'Components/sd-link',
   component: 'sd-link',
   args: overrideArgs([
-    { type: 'slot', name: 'default', value: 'Link' },
+    { type: 'slot', name: 'default', value: 'Default' },
     { type: 'attribute', name: 'href', value: '#' }
   ]),
   argTypes,
@@ -32,297 +39,119 @@ export const Default = {
 };
 
 /**
- * Unset `href` to make the link disabled.
+ * Use the `standalone` attribute to control the layout of the icon and text within the component.
+ * If true, the icon and text will be displayed side by side, each occupying its own column.
+ * If false or not provided, the icon will be displayed inline within the text.
  */
 
-export const InvertedAndDisabled = {
-  name: 'Disabled × Inverted',
-  parameters: { controls: { exclude: ['size', 'default', 'href', 'inverted'] } },
-  render: (args: any) =>
-    generateTemplate({
-      axis: {
-        x: {
-          type: 'attribute',
-          name: 'href',
-          values: ['#link', { value: '', title: '–' }]
-        },
-        y: {
-          type: 'attribute',
-          name: 'inverted',
-          values: [false, true]
-        }
-      },
-      args,
-      options: {
-        templateBackgrounds: { alternate: 'y', colors: ['white', 'rgb(var(--sd-color-primary, 0 53 142))'] }
-      }
-    })
+export const Standalone = {
+  name: 'Standalone',
+  render: () => html`
+    <div class="flex gap-12">
+      <div class="sd-prose">
+        <ul>
+          <li>
+            <sd-link href="http://union-investment.com" standalone>
+              <sd-icon library="global-resources" name="system/home" slot="icon-left"></sd-icon>
+              Union Investment
+            </sd-link>
+          </li>
+          <li>
+            <sd-link href="https://solid-design-system.fe.union-investment.de/x.x.x/storybook/" standalone>
+              <sd-icon library="global-resources" name="system/pen" slot="icon-left"></sd-icon>
+              Solid Design System
+            </sd-link>
+          </li>
+        </ul>
+
+        <p>
+          While the list above shows standalone links, we now will link to the
+          <sd-link href="https://cd.union-investment.de">CD Toolbox</sd-link> inside a paragraph.
+        </p>
+      </div>
+    </div>
+  `
 };
 
 /**
- * You can make links bold by setting `<b>` tags around the text in the main slot.
+ * Use the `size` attribute to adjust the size of the link. By default it is set to `inherit` to adapt to the surrounding text but can be changed to `lg` or `sm`.
  */
 
-export const BoldInMainSlot = {
-  name: 'Bold',
-  parameters: { controls: { exclude: ['default'] } },
-  render: (args: any) =>
-    generateTemplate({
-      axis: {
-        x: {
-          type: 'slot',
-          name: 'default',
-          title: 'main-slot',
-          values: ['Link', '<b>Link</b>']
-        }
-      },
-      args,
-      options: {
-        templateBackgrounds: { alternate: 'y', colors: ['white', 'rgb(var(--sd-color-primary, 0 53 142))'] }
-      }
-    })
+export const Size = {
+  name: 'Size',
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-link href="https://solid-design-system.fe.union-investment.de/x.x.x/storybook/">Inherit</sd-link>
+      <sd-link href="https://solid-design-system.fe.union-investment.de/x.x.x/storybook/" size="lg">Large</sd-link>
+      <sd-link href="https://solid-design-system.fe.union-investment.de/x.x.x/storybook/" size="sm">Small</sd-link>
+    </div>
+  `
 };
 
 /**
- * Use the `icon-left` and `icon-right` slots to add icons. They automatically adapt the size.
+ * Use the `inverted` attribute when displayed on primary background.
  */
 
-export const SizeAndIconSlots = {
-  name: 'Size × Icon Slots',
-  parameters: { controls: { exclude: ['default', 'icon-left', 'size', 'icon-right', 'standalone'] } },
-  render: (args: any) => {
-    return html`
-      ${[false, true].map(standalone =>
-        generateTemplate({
-          axis: {
-            x: {
-              // To make the story creation easier, we put everything in the default slot.
-              type: 'slot',
-              name: 'default',
-              title: 'slot="icon-..."',
-              values: [
-                { value: 'Link', title: '–' },
-                {
-                  value: 'Link<sd-icon library="global-resources" name="system/picture" slot="icon-left"></sd-icon>',
-                  title: 'left'
-                },
-                {
-                  value: 'Link<sd-icon library="global-resources" name="system/picture" slot="icon-right"></sd-icon>',
-                  title: 'right'
-                }
-              ]
-            },
-            y: { type: 'attribute', name: 'size' }
-          },
-          options: { title: `standalone="${standalone ? 'true' : 'false'}"` },
-          constants: { type: 'attribute', name: 'standalone', value: standalone },
-          args
-        })
-      )}
-    `;
-  }
+export const Inverted = {
+  name: 'Inverted',
+  render: () => html`
+    <div class="flex gap-12 bg-primary p-4">
+      <sd-link href="https://solid-design-system.fe.union-investment.de/x.x.x/storybook/" inverted>Inverted</sd-link>
+    </div>
+  `
 };
 
 /**
- * Icons automatically adapt to the link's invertedness.
+ * Remove the `href` to disable the link.
  */
 
-export const InvertedAndIconSlots = {
-  name: 'Inverted × Icon Slots',
-  parameters: { controls: { exclude: ['inverted', 'default', 'icon-left', 'icon-right'] } },
-  render: (args: any) =>
-    generateTemplate({
-      axis: {
-        // To make the story creation easier, we put everything in the default slot.
-        x: {
-          type: 'slot',
-          name: 'default',
-          title: 'slot="icon-..."',
-          values: [
-            { value: 'Link', title: '–' },
-            {
-              value: 'Link<sd-icon library="global-resources" name="system/picture" slot="icon-left"></sd-icon>',
-              title: 'left'
-            },
-            {
-              value: 'Link<sd-icon library="global-resources" name="system/picture" slot="icon-right"></sd-icon>',
-              title: 'right'
-            }
-          ]
-        },
-        y: {
-          type: 'attribute',
-          name: 'inverted',
-          values: [false, true]
-        }
-      },
-      args,
-      options: {
-        templateBackgrounds: { alternate: 'y', colors: ['white', 'rgb(var(--sd-color-primary, 0 53 142))'] }
-      }
-    })
+export const Disabled = {
+  name: 'Disabled',
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-link>Disabled</sd-link>
+    </div>
+  `
 };
 
 /**
- * This `standalone` attribute controls the layout of the icon and text within the component.
- * - If true, the icon and text will be displayed side by side, each occupying its own column.
- * - If false or not provided, the icon will be displayed inline within the text.
+ * Use the `href` attribute to enable the link.
+ *
+ * Use the `target` attribute to specify where to open the link.
+ *
+ * Use the `download` attribute to tell the browser to download the linked file as this filename.
  */
 
-export const StandaloneAndIconSlots = {
-  name: 'Standalone × Icon Slots',
-  parameters: { controls: { exclude: ['icon-right', 'icon-left', 'main', 'standalone', 'default'] } },
-  render: (args: any) => {
-    return html`
-      ${[false, true].map(surroundingContent =>
-        // To make the story creation easier, we put everything in the default slot.
-        generateTemplate({
-          axis: {
-            x: {
-              type: 'slot',
-              name: 'default',
-              title: 'icon-...',
-              values: [
-                { value: 'Magna ex ex elit cupidatat non esse.', title: '–' },
-                {
-                  value:
-                    '<sd-icon library="global-resources" name="system/picture" slot="icon-left"></sd-icon>Magna ex ex elit cupidatat non esse.',
-                  title: 'left'
-                },
-                {
-                  value:
-                    'Magna ex ex elit cupidatat non esse.<sd-icon library="global-resources" name="system/picture" slot="icon-right"></sd-icon>',
-                  title: 'right'
-                }
-              ]
-            },
-            y: {
-              type: 'attribute',
-              name: 'standalone',
-              values: [false, true]
-            }
-          },
-          constants: [
-            {
-              type: 'template',
-              name: 'default-template',
-              value: `<div style="font-size: 16px; text-align: left; width: 200px; word-break: break-all;">${
-                surroundingContent ? 'Qui do.' : ''
-              }%TEMPLATE%${surroundingContent ? 'Eiusmod minim excepteur.</div>' : ''}`
-            }
-          ],
-          args,
-          options: {
-            title: surroundingContent ? 'example with surrounding content' : 'example without surrounding content'
-          }
-        })
-      )}
-    `;
-  }
+export const AsLink = {
+  name: 'As link',
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-link
+        href="https://solid-design-system.fe.union-investment.de/x.x.x/storybook/?path=/docs/docs-general-introduction--docs"
+        >Link</sd-link
+      ><sd-link href="https://union-investment.com" target="_blank">New Window</sd-link
+      ><sd-link href="./placeholders/src/images/collaboration.jpg" download>Download</sd-link>
+    </div>
+  `
 };
 
 /**
- * Use the part selector to align the icon within the component.
+ * Use the `icon-left` and `icon-right` slots to add system icons to each side of the link. They automatically adapt the size.
  */
 
-export const IconAlignment = {
-  parameters: {
-    controls: { exclude: ['icon-right', 'default', 'standalone', 'base'] }
-  },
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        y: {
-          type: 'template',
-          name: 'sd-link::part(base){align-items: ...;}',
-          values: ['start', 'center', 'end'].map(alignIcon => {
-            return {
-              value: `<style>#align-icon-${alignIcon} sd-link::part(base){align-items: ${alignIcon};}</style><div id="align-icon-${alignIcon}">%TEMPLATE%</div>`,
-              title: alignIcon
-            };
-          })
-        }
-      },
-      constants: [
-        {
-          type: 'template',
-          name: 'constant-template',
-          value: '<style>div[id^=align-icon-]{text-align: left; width: 300px;}</style>'
-        },
-        {
-          type: 'slot',
-          name: 'icon-right',
-          value: '<sd-icon library="global-resources" name="system/arrow-right" slot="icon-right"></sd-icon>'
-        },
-        {
-          type: 'slot',
-          name: 'default',
-          value: 'In dolore consectetur do excepteur tempor occaecat magna anim esse sit dolor mollit est voluptate.'
-        },
-        {
-          type: 'attribute',
-          name: 'standalone',
-          value: true
-        }
-      ],
-      args
-    });
-  }
-};
+export const Icon = {
+  name: 'Icon',
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-link href="https://solid-design-system.fe.union-investment.de/x.x.x/storybook/" standalone>
+        <sd-icon library="global-resources" name="system/picture" slot="icon-left"></sd-icon>
+        Icon Left
+      </sd-link>
 
-/**
- * Use the `base`, `label`, `icon-left` and `icon-right` part selectors to customize the button.
- */
-
-export const Parts = {
-  parameters: {
-    controls: { exclude: ['base', 'label', 'icon-left', 'icon-right'] }
-  },
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        y: {
-          type: 'template',
-          name: 'sd-link::part(...){outline: solid 2px red}',
-          values: ['base', 'label', 'icon-left', 'icon-right'].map(part => {
-            return {
-              title: part,
-              value: `<style>#part-${part} sd-link::part(${part}){outline: solid 2px red}</style><div id="part-${part}">%TEMPLATE%</div>`
-            };
-          })
-        }
-      },
-      constants: [
-        {
-          type: 'slot',
-          name: 'icon-right',
-          value: '<sd-icon library="global-resources" name="system/picture" slot="icon-right"></sd-icon>'
-        },
-        {
-          type: 'slot',
-          name: 'icon-left',
-          value: '<sd-icon library="global-resources" name="system/picture" slot="icon-left"></sd-icon>'
-        }
-      ],
-      args
-    });
-  }
-};
-
-/**
- * sd-links are fully accessibile via keyboard.
- */
-
-export const Mouseless = {
-  render: (args: any) => {
-    return html`<div class="mouseless">${generateTemplate({ args })}</div>`;
-  },
-
-  play: async ({ canvasElement }: { canvasElement: HTMLUnknownElement }) => {
-    const el = canvasElement.querySelector('.mouseless sd-link');
-    await waitUntil(() => el?.shadowRoot?.querySelector('a'));
-    // We have to catch the event as otherwise Storybook will break
-    el!.addEventListener('click', e => e.preventDefault());
-    await userEvent.type(el!.shadowRoot!.querySelector('a')!, '{return}', { pointerEventsCheck: 0 });
-  }
+      <sd-link href="https://solid-design-system.fe.union-investment.de/x.x.x/storybook/" standalone>
+        Icon Right
+        <sd-icon library="global-resources" name="system/picture" slot="icon-right"></sd-icon>
+      </sd-link>
+    </div>
+  `
 };
